[glib2] Update to 2.56.1. Contributes to JB#41420
- drop syslog patch, glib supports structured logging now
- mtab patch could have been applied, but the rationale for inclusion
in the first place is not available with this git repo. Just omitted in
hope of getting rid of it. Upstream by default already moved to libmount,
but that seems to have problems  with Tracker. Enabling commit
(496cd8aac6402) does state that it returns slightly different results,
namely with bind mounts.